Port strat_correct to Fortran
