+++ docs/stratification_correction/solve.f
@@ -1,0 +1,9 @@
+      program solve
+         real z0, z1, L, Ri_b
+
+         z0 = 0.01
+         z1 = 1.
+         L = 1.
+         Ri_b = 0.14
+         print *, psi_m_stable(0.05 / 0.1)
+      end program
