+++ docs/stratification_correction/solve.py
@@ -1,77 +1,18 @@
-from functools import partial
-
 import numpy as np
+
+# Fortran
+# Note: execute `make lib` to generate the Fortran library
+import strat_correct_f as ftn
 from scipy.optimize import newton
 
-
-def root2_1m16(xi):
-    return (1.0 - 16 * xi) ** 0.5
-
-
-def root4_1m16(xi):
-    return (1.0 - 16 * xi) ** 0.25
+# Python
+# from strat_correct import f_richardson, df_richardson_dL
 
 
-def psi_m_stable(xi):
-    return -5 * xi
+f_richardson = np.vectorize(ftn.f_richardson)
+df_richardson_dL = np.vectorize(ftn.df_richardson_dl)
 
-
-def eval_func_at(z0, z1, L, func):
-    r"""Evaluate a integrated stability function :math:`\psi` at a
-    elevation :math:`z_1` with roughness :math:`z_0` and
-    Obukhov length :math:`L`.
-    """
-    return np.log(z1 / z0) - func(z1 / L) + func(z0 / L)
-
-
-PsiM_stable_at = partial(eval_func_at, func=psi_m_stable)
-# Because psi_h_stable is identical to psi_m_stable
-PsiH_stable_at = PsiM_stable_at
-
-
-def psi_m_unstable(xi):
-    """See ``psi_m_unstable_param`` in the notebook."""
-    xi4 = root4_1m16(xi)
-    return (
-        np.log((1 / 8) * (xi4 + 1) ** 2 * (xi4 ** 2 + 1))
-        - 2 * np.arctan(xi4)
-        + (1 / 2) * np.pi
-    )
-
-
-def psi_h_unstable(xi):
-    """See ``psi_h_unstable_param`` in the notebook."""
-    xi2 = root2_1m16(xi)
-    return np.log((1 / 4) * (xi2 + 1) ** 2)
-
-
-PsiM_unstable_at = partial(eval_func_at, func=psi_m_unstable)
-PsiH_unstable_at = partial(eval_func_at, func=psi_h_unstable)
-
-
-@np.vectorize
-def richardson(L, z0, z1):
-    if L >= 0:
-        PsiM_at = PsiM_stable_at
-        PsiH_at = PsiH_stable_at
-    else:
-        PsiM_at = PsiM_unstable_at
-        PsiH_at = PsiH_unstable_at
-
-    # For Dirichlet conditions
-    return (z1 / L) * (PsiH_at(z0, z1, L) / PsiM_at(z0, z1, L) ** 2)
-
-
-def f_richardson(L, z0, z1, Ri_b):
-    return Ri_b - richardson(L, z0, z1)
-
-
-def df_richardson_dL(L, z0, z1, Ri_b):
-    """Numerical first derivative by Central difference."""
-    dL = 1e-4 * L
-    f_plus = f_richardson(L + dL, z0, z1, Ri_b)
-    f_minus = f_richardson(L - dL, z0, z1, Ri_b)
-    return (f_plus - f_minus) / (2 * dL)
+TOLERANCE = 1e-8
 
 
 def initial_guess_L(z0, Ri_b):
@@ -94,7 +35,12 @@ def solve_obukhov_len_secant(z0, z1, Ri_b, L0):
     try:
         # NOTE: No fprime parameter provided
         root, result = newton(
-            f_richardson, L0, args=(z0, z1, Ri_b), maxiter=1000, full_output=True
+            f_richardson,
+            L0,
+            args=(z0, z1, Ri_b),
+            tol=TOLERANCE,
+            maxiter=1000,
+            full_output=True,
         )
     except RuntimeError:
         print(f"{Ri_b=}, {L0=}")
@@ -120,6 +66,7 @@ def solve_obukhov_len_central(z0, z1, Ri_b, L0):
             f_richardson,
             L0,
             args=(z0, z1, Ri_b),
+            tol=TOLERANCE,
             maxiter=1000,
             full_output=True,
             fprime=df_richardson_dL,
@@ -150,8 +97,13 @@ if __name__ == "__main__":
     import matplotlib.pyplot as plt
     import numpy as np
 
-    Ri_small = 1e-2
-    Ri_bs = np.concatenate((np.linspace(-5, -Ri_small), np.linspace(Ri_small, 0.22)))
+    Ri_small = 0.01
+    Ri_bs = np.concatenate(
+        (
+            np.linspace(-5, -Ri_small),
+            np.linspace(Ri_small, 0.22),
+        )
+    )
 
     z0 = 0.1
     z1 = 10 * z0
@@ -200,32 +152,32 @@ if __name__ == "__main__":
     fig.suptitle(f"{z0=} {z1=}")
     fig.tight_layout()
     plt.show()
-# -
+    # -
 
-# Ensure the sign of L is correct
+    # Ensure the sign of L is correct
 
-np.all(np.sign(Ri_bs) == np.sign(Ls))
+    np.all(np.sign(Ri_bs) == np.sign(Ls))
 
-# +
-L_thresh = 5
+    # +
+    L_thresh = 5
 
-RI_B, L = np.meshgrid(Ri_bs, np.linspace(-L_thresh, L_thresh))
-F = f_richardson(L, z0, z1, RI_B)
-DF_DL = df_richardson_dL(L, z0, z1, RI_B)
+    RI_B, L = np.meshgrid(Ri_bs, np.linspace(-L_thresh, L_thresh))
+    F = f_richardson(L, z0, z1, RI_B)
+    DF_DL = df_richardson_dL(L, z0, z1, RI_B)
 
-fig, axes = plt.subplots(ncols=3, sharey=True, figsize=(15, 4))
-ax0, ax1, ax2 = axes.ravel()
-fig.colorbar(ax0.contourf(RI_B, L, F), ax=ax0)
-fig.colorbar(ax1.contourf(RI_B, L, DF_DL, levels=20, cmap="inferno"), ax=ax1)
-fig.colorbar(ax2.contourf(RI_B, L, np.log(abs(DF_DL)), cmap="inferno"), ax=ax2)
+    fig, axes = plt.subplots(ncols=3, sharey=True, figsize=(15, 4))
+    ax0, ax1, ax2 = axes.ravel()
+    fig.colorbar(ax0.contourf(RI_B, L, F), ax=ax0)
+    fig.colorbar(ax1.contourf(RI_B, L, DF_DL, levels=20, cmap="inferno"), ax=ax1)
+    fig.colorbar(ax2.contourf(RI_B, L, np.log(abs(DF_DL)), cmap="inferno"), ax=ax2)
 
-ax0.plot(Ri_bs, L0, color="red", label="Initial $L_0$")
-ax0.plot(Ri_bs, Ls, color="white", label="Final $L$")
-ax0.set_ylim(-L_thresh, L_thresh)
-ax0.legend()
+    ax0.plot(Ri_bs, L0, color="red", label="Initial $L_0$")
+    ax0.plot(Ri_bs, Ls, color="white", label="Final $L$")
+    ax0.set_ylim(-L_thresh, L_thresh)
+    ax0.legend()
 
-for ax, title in zip(
-    axes, ("$f$", r"$\partial f/ \partial L$", r"$\ln(|\partial f / \partial L|)$")
-):
-    ax.set(ylabel="L", xlabel="$Ri_b$", title=title)
-# -
+    for ax, title in zip(
+        axes, ("$f$", r"$\partial f/ \partial L$", r"$\ln(|\partial f / \partial L|)$")
+    ):
+        ax.set(ylabel="L", xlabel="$Ri_b$", title=title)
+    # -
